feat(): add observability services with Grafana Cloud (#17)
Introducing the Grafana Cloud agents to the GKE clusters. This follows
the official Grafana Cloud documentation so we can have:
1. Prometheus metrics for Kubernetes code components (cadvisor, nodes,
api server, events, etc...)
2. Logs from all the containers running on our clusters

We use the same environment variables model to pass Grafana Cloud
passwords & URLs to Terraform, which stores those variables in a
`ConfigMap`, used by the Flux `kustomize-controller` to do variable
substitution.

> This doesn't enable the monitoring of the services already present in
the cluster (Traefik & CertManager). It will be done in another PR