feat(): self signed certs
Changes the content of the `kube-security` namespace with a new service:
trust-manager

This allow us to create our own Certificate Authority with Cert Manager,
and bundle it with other public CA into a single ca.crt.

We can then distribute this CA Cert to PODs so they see self signed
certificate as valid.